Switch signature of *addch to take int's instead of uint8's
Apparently ncurses' chtype is an int (to fit in all the extra attribute
bits and such) and it doesn't work correctly if the values are
truncated. orz.

+++ curses.go
@@ -154,11 +154,11 @@ func Mvaddstr(y, x int, str ...interface{}) {
     C.mvaddstr(C.int(y), C.int(x), res)
 }
 
-func Addch(ch uint8) {
+func Addch(ch int) {
 	C.addch(C.chtype(ch))
 }
 
-func Mvaddch(y, x int, ch uint8) {
+func Mvaddch(y, x int, ch int) {
 	C.mvaddch(C.int(y), C.int(x), C.chtype(ch))
 }
 
@@ -174,11 +174,11 @@ func (win *Window) Mvaddstr(y, x int, str ...interface{}) {
     C.mvwaddstr((*C.WINDOW)(win), C.int(y), C.int(x), res)
 }
 
-func (win *Window) Addch(ch uint8) {
+func (win *Window) Addch(ch int) {
 	C.waddch((*C.WINDOW)(win), C.chtype(ch))
 }
 
-func (win *Window) Mvaddch(y, x int, ch uint8) {
+func (win *Window) Mvaddch(y, x int, ch int) {
 	C.mvwaddch((*C.WINDOW)(win), C.int(y), C.int(x), C.chtype(ch))
 }
